refactor(structured): add property list renderer
